+++ src/bin/selection.rs
@@ -36,10 +36,7 @@ struct Args {
 }
 
 fn main() -> Result<(), EventLoopError> {
-    if std::env::var("RUST_LOG").is_err() {
-        std::env::set_var("RUST_LOG", "info");
-    }
-    env_logger::init();
+    env_logger::Builder::from_env(env_logger::Env::default().default_filter_or("info")).init();
 
     let event_loop = EventLoop::new()?;
     event_loop.run_app(&mut App::new(Args::parse()))?;