+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 ### Added
 
 - 🔢 Multi-model viewer to see multiple Gaussian models, enabled by `multi-model` feature.
+- 🎥 Add depth stencil as an option to render with.
 
 ### Changed
 
+- ✅ Update to Rust 2024 edition.
 - ⏫ Increase the capability of the viewer to handle larger models.
 
 ## [0.1.0](https://crates.io/crates/wgpu-3dgs-viewer/0.1.0) - 2025-02-27